Added support for cdn prefix and optional random numbered cdn cname.  Updated test suite with new tests and  leveraging Jasmine.
