+++ package.json
@@ -35,7 +35,9 @@
     "collector"
   ],
   "dependencies": {
+    "assert": "^1.3.0",
     "gulp-util": "^3.0.0",
+    "path": "^0.11.14",
     "through2": ">=0.5.1",
     "underscore": "^1.6.0"
   },
